[Travis] Do not install a custom libc++ and libc++abi
I must have done that because libstdc++ was not sufficiently close to
modern standards at the time, but the right thing to do is to rely on
the Standard Library provided by the system and/or the compiler.
Otherwise we're always testing against the same standard library,
which is not realistic.